Zooka_app/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-cloud-spanner==3.40.1
werkzeug==3.0.1
google-cloud-aiplatform==1.130.0